Revert "Migrate users of mysten_metrics Histogram to prometheus Histogram (#19043)"
This reverts commit ed69f7cce7ac200a849647c5d8efa7113969d44a.

+++ crates/sui-oracle/src/lib.rs
@@ -525,7 +525,7 @@ impl OnChainDataUploader {
                 self.metrics
                     .uploaded_values
                     .with_label_values(&[&data_point.feed_name])
-                    .observe(data_point.value as f64);
+                    .observe(data_point.value);
             } else {
                 self.metrics
                     .upload_data_errors
@@ -538,9 +538,7 @@ impl OnChainDataUploader {
         let storage_rebate = effects.gas_cost_summary().storage_rebate;
         let computation_cost = effects.gas_cost_summary().computation_cost;
         let net_gas_usage = effects.gas_cost_summary().net_gas_usage();
-        self.metrics
-            .computation_gas_used
-            .observe(computation_cost as f64);
+        self.metrics.computation_gas_used.observe(computation_cost);
         self.metrics.total_gas_cost.inc_by(gas_usage);
         self.metrics.total_gas_rebate.inc_by(storage_rebate);
 
@@ -662,7 +660,7 @@ impl OnChainDataReader {
                         self.metrics
                             .downloaded_values
                             .with_label_values(&[feed_name])
-                            .observe(value * METRICS_MULTIPLIER);
+                            .observe((value * METRICS_MULTIPLIER) as u64);
                         self.metrics
                             .download_successes
                             .with_label_values(&[feed_name, &object_id.to_string()])
